+++ api/app/main.py
@@ -102,7 +102,9 @@ from .obs import capture_exception, init_sentry
 from .otel import init_tracing
 from .routes_admin_menu import router as admin_menu_router
 from .routes_admin_privacy import router as admin_privacy_router
+
 from .routes_admin_qrpack import router as admin_qrpack_router
+
 from .routes_alerts import router as alerts_router
 from .routes_api_keys import router as api_keys_router
 from .routes_auth_2fa import router as auth_2fa_router
@@ -137,6 +139,7 @@ from .routes_menu_import import router as menu_import_router
 from .routes_metrics import router as metrics_router
 from .routes_metrics import ws_messages_total
 from .routes_onboarding import router as onboarding_router
+from .routes_order_void import router as order_void_router
 from .routes_orders_batch import router as orders_batch_router
 from .routes_outbox_admin import router as outbox_admin_router
 from .routes_owner_aggregate import router as owner_aggregate_router
@@ -151,9 +154,11 @@ from .routes_pwa_version import router as pwa_version_router
 from .routes_qrpack import router as qrpack_router
 from .routes_ready import router as ready_router
 from .routes_reports import router as reports_router
+from .routes_csp import router as csp_router
 from .routes_security import router as security_router
 from .routes_staff import router as staff_router
 from .routes_support import router as support_router
+from .routes_admin_support import router as admin_support_router
 from .routes_support_bundle import router as support_bundle_router
 from .routes_tables_map import router as tables_map_router
 from .routes_tables_qr_rotate import router as tables_qr_rotate_router
@@ -833,6 +838,7 @@ app.include_router(hotel_guest_router)
 app.include_router(invoice_pdf_router)
 app.include_router(onboarding_router)
 app.include_router(qrpack_router)
+app.include_router(order_void_router)
 
 # KDS/KOT domain
 app.include_router(kot_router)
@@ -869,6 +875,7 @@ app.include_router(version_router)
 app.include_router(ready_router)
 app.include_router(help_router)
 app.include_router(support_router)
+app.include_router(admin_support_router)
 app.include_router(support_bundle_router)
 app.include_router(legal_router)
 app.include_router(maintenance_router)
@@ -890,6 +897,7 @@ app.include_router(admin_qrpack_router)
 app.include_router(daybook_pdf_router)
 app.include_router(digest_router)
 app.include_router(reports_router)
+app.include_router(csp_router)
 app.include_router(gst_monthly_router)
 app.include_router(exports_router)
 app.include_router(export_all_router)